Factor getData out into a library method.
