+++ frontend/src/home/Stories.vue
@@ -3,25 +3,14 @@ import StoryCard from '@/components/StoryCard.vue'
 
 import { ref, onMounted } from 'vue'
 import { useRouter } from 'vue-router';
-import Papa from 'papaparse';
+import { getData } from '@/components/fetchSpreadsheet'
 const sid = import.meta.env.VITE_WOMEN_CONTENT_SHEET_ID
 
 const StoriesURI = `https://docs.google.com/spreadsheets/d/${sid}/export?format=csv`
 const rows = ref([]);
 
-async function getData() {
-  await Papa.parse(StoriesURI, {
-    header: true,
-    download: true,
-    worker: true,
-    complete: function (results, file) {
-      rows.value = results.data;
-    },
-  });
-}
-
 onMounted(() => {
-  getData();
+  getData(StoriesURI, rows);
 });
 
 const router = useRouter();
